+++ 3rdparty/MelodixAPI/client/MDGetTopPlaylist_200_response_playlists_inner.h
@@ -48,8 +48,8 @@ public:
     bool is_name_Set() const;
     bool is_name_Valid() const;
 
-    qint32 getId() const;
-    void setId(const qint32 &id);
+    double getId() const;
+    void setId(const double &id);
     bool is_id_Set() const;
     bool is_id_Valid() const;
 
@@ -218,7 +218,7 @@ private:
     bool m_name_isSet;
     bool m_name_isValid;
 
-    qint32 m_id;
+    double m_id;
     bool m_id_isSet;
     bool m_id_isValid;
